Fix filesystem setup

+++ scripts/generate_calculated_paths_env_file.sh
@@ -18,6 +18,8 @@ if [ -z "$APP_IS_EXPOSED" ]; then
     echo "APP_IS_EXPOSED must be set" >&2
     exit 1
 fi
+APP_IS_EXPOSED=$(echo "$APP_IS_EXPOSED" | tr '[:upper:]' '[:lower:]')
+
 
 if [ $APP_IS_EXPOSED != "true" ] && [ $APP_IS_EXPOSED != "false" ]; then
     echo "APP_IS_EXPOSED must be set to true or false" >&2
@@ -39,10 +41,24 @@ else
     FLASK_LOG_DIR=${BASE_DIR}${FLASK_LOG_INTERNAL_DIR}
 fi
 
+required_vars=("FLASK_LOGS_ARE_NEEDED")
+if [ -z "$FLASK_LOGS_ARE_NEEDED" ]; then
+    echo "FLASK_LOGS_ARE_NEEDED must be set" >&2
+    exit 1
+fi
+FLASK_LOGS_ARE_NEEDED=$(echo "$FLASK_LOGS_ARE_NEEDED" | tr '[:upper:]' '[:lower:]')
+
+if [ $FLASK_LOGS_ARE_NEEDED != "true" ] && [ $FLASK_LOGS_ARE_NEEDED != "false" ]; then
+    echo "FLASK_LOGS_ARE_NEEDED must be set to true or false."
+    exit 1
+fi
+
+if [ $FLASK_LOGS_ARE_NEEDED = "true"]; then
+    echo "FLASK_LOG_DIR: $FLASK_LOG_DIR"
+    echo "FLASK_LOG_DIR=$FLASK_LOG_DIR" >> "$GENERATED_PATHS_ENV_FILE"
+fi
+
 if [ -f $GENERATED_PATHS_ENV_FILE ]; then
     rm -f $GENERATED_PATHS_ENV_FILE
 fi
 touch $GENERATED_PATHS_ENV_FILE
-
-echo "FLASK_LOG_DIR: $FLASK_LOG_DIR"
-echo "FLASK_LOG_DIR=$FLASK_LOG_DIR" >> "$GENERATED_PATHS_ENV_FILE"
